Add the ability to read from a file using Stream
Why use Stream:

Once in a while, there will be problems which could be solved with list
and transformations, except that the list is big and consumes lot of
memory. Let’s say you want to export records in database as a csv
report, it might not be possible to load the entire set of records in
memory before the transformations can be applied to it. Stream
abstraction in Elixir is designed to address these kinds of problems.

The stream abstraction helps us to apply a sequence of such
transformations on a stream with a constant memory overhead.

Also, The stream implements both Enumerable and Collectable protocols,
which means it can be used both for read and write. This is specified
under File.stream! docs

source: http://ananthakumaran.in/2017/11/28/stream.html