Refactor the logic that helps determine what sales tax is applicable for the item
If an item is imported and whether or not it has basic sales tax
applicable is something that is determined by checking item specific
properties and hence we’re placing the appropriate checks for the same
in the item module.

Since we’re parsing the other line item details quantity, product,
price with their appropriate types, we could do the same to update if a
given item is imported and whether or not basic sales tax is applicable
for that item.